enabled range endpoint, minor refactoring

+++ Application/Common/Models/BitcoinPricesRangeDto.cs
@@ -5,7 +5,7 @@ namespace Application.Common.Models
 {
     public class BitcoinPricesRangeDto
     {
-        public Dictionary<DateTime, double> Bpi { get; set; }
+        public Dictionary<DateTime, float> Bpi { get; set; }
         public string Disclaimer { get; set; }
         public BitcoinPriceTimeDto Time { get; set; }
     }
